refactor simulation code
